Requirement ID, add formatting to tooltip also
